Move Modal Components from HomeScreen to SimulatorScreen
